+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### `@liveblocks/react-ui`
 
+- Add `gap` prop to `AvatarStack` to control the `--lb-avatar-stack-gap` CSS
+  variable.
+- Add `padding` prop to `CommentPin` to control the `--lb-comment-pin-padding`
+  CSS variable.
 - Fix `size` props on `AvatarStack` and `CommentPin` not working as expected
   when passing numbers.
 - Fix `autoFocus` prop on `FloatingComposer`.